Consolidate into a single service and update docs
