+++ consensus/src/terminator.rs
@@ -166,22 +166,18 @@ impl Terminator {
     }
 }
 
-pub async fn handle_task_termination<T>(
-    task_handle: T,
-    target: &'static str,
-    name: &'static str,
-    index: aleph_bft_types::NodeIndex,
-) where
+pub async fn handle_task_termination<T>(task_handle: T, target: &'static str, name: &'static str)
+where
     T: FusedFuture<Output = Result<(), ()>>,
 {
     if !task_handle.is_terminated() {
         if let Err(()) = task_handle.await {
             warn!(
                 target: target,
-                "{:?} {} task stopped with an error", index, name
+                "{} task stopped with an error", name
             );
         }
-        debug!(target: target, "{:?} {} stopped.", index, name);
+        debug!(target: target, "{} stopped.", name);
     }
 }
